opentsdb: support Gexp group by tags (#7118)
* support Gexp group by tags
* support Gexp alias
* change to use Gexp query as display metric

The query merge of gexp was removed,
in order to make the response data could be identified.
Since only the metric and tags are shown in response data,
there is not enough to identify the gexp query.

The modifications are based on pull request (#7118),
which addresses the feature of request (#3805).